requirements.txt
streamlit
numpy
pandas
plotly
scikit-learn
google-generativeai
openai